+++ R/trans_venn.R
@@ -35,8 +35,9 @@
 #' @return Returns a Venn plot by grouping variable(s).
 #' @section Warning: The algorithm used to overlap the Venn circles becomes 
 #' increasingly overburdened and less accurate with increased grouping 
-#' variables. An alternative is to use a network plot with dissimilarity 
-#' measures labeling the edges between nodes (grouping variables).
+#' variables. An alternative is to use a network plot with 
+#' \{code{\link[qdap]{Dissimilarity}} measures labeling the edges between nodes 
+#' (grouping variables) or a heat map (\code{\link[qdap]{qheat}}).
 #' @seealso \code{\link[venneuler]{venneuler}}
 #' @keywords venn
 #' @export